Update to use api.v2.emissions-api.org
This patch updates the website to use the new version of the emissions api.
The deck.gl example has been changed to show now the US instead of Germany,
since with the new hexagons it is possible to use a broader area and Germany
looks a little bit tiny.

The roadmap has been deleted, since it is outdated for now and even if we
update it, it will soon be outdated again, when we are working on this.

The dependencies and the .gitignore are also updated.